requirements/base.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-appconf==1.0.3
django-betterforms==1.2
django-debug-toolbar==2.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
drf-nested-routers==0.91
dropbox==9.4.0
idna==2.8
markdown2==2.3.8
psycopg2-binary==2.8.3
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
URLObject==2.4.3
